refac: Move cobra commands files to new cli package
Leave main.go in cmd/
