+++ cli/testdata/down_tests/case_1/docker-compose-scripts/.env
