+++ cli/testdata/logs_tests/case_1/config/config.yaml
