+++ cli/testdata/logs_tests/case_1/docker-compose-scripts/.env
