+++ cli/testdata/logs_tests/case_1/docker-compose-scripts/docker-compose.yml
