+++ cli/testdata/root_test/config/config.yaml
