+++ cli/version_test.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"bytes"
