--- cmd/1click/main.go
@@ -1,11 +1,0 @@
-/*
-Copyright © 2022 Nethermind hello.nethermind.io
-
-*/
-package main
-
-import "github.com/NethermindEth/1click/cmd"
-
-func main() {
-	cmd.Execute()
-}
