--- cmd/version.go
@@ -1,25 +1,0 @@
-/*
-Copyright © 2022 Nethermind hello.nethermind.io
-
-*/
-package cmd
-
-import (
-	"fmt"
-
-	"github.com/spf13/cobra"
-)
-
-// versionCmd represents the version command
-var versionCmd = &cobra.Command{
-	Use:   "version",
-	Short: "print 1click version",
-	Long:  ``,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Fprintln(cmd.OutOrStdout(), "1click 0.1")
-	},
-}
-
-func init() {
-	rootCmd.AddCommand(versionCmd)
-}
